requirements.txt
pycrypto
sqlalchemy==1.2.15
requests
Flask==1.0.2
mysqlclient
mysql-connector-python